Engine class, entity class, map and tiles classes, refactoring for main and actions
Added an Engine class where appropriate method calls are made to in game events, and manages updates and changes to the console output.

Added an Entity class that generically defines in game objects and actors. Simplifies the rendering process and event resolution in engine.

Added map and tile type classes to categorize and classify types of tiles (and their interactions with entities) for dungeon generation, and the map holds the level layout.